Backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.1
python-dotenv==1.0.0
alembic==1.13.1
authlib==1.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
google-generativeai==0.8.5
pgvector==0.4.1
databases==0.9.0
aiofiles==24.1.0
boto3==1.35.0
httpx==0.27.0
requests==2.31.0
supabase==2.9.1
python-multipart==0.0.6
Pillow==10.2.0
websockets==15.0.1
numpy==1.26.2
scikit-learn==1.3.2
email-validator==2.1.0.post1
pypdf==3.17.4
PyPDF2==3.0.1
beautifulsoup4==4.12.3
google-api-python-client==2.120.0